backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
pydantic>=2.0
pydantic-settings>=2.0
pyyaml
python-dotenv
greenlet
openai # For LLM calls
python-docx # Since we already used it, might as well keep it if we need to parse uploaded docs later.
aiosqlite
pydantic[email]

# 认证相关
python-jose[cryptography]  # JWT
passlib[bcrypt]  # 密码加密
bcrypt  # bcrypt 加密算法
